+++ search_backend/api/lib/search_formatting_functions.py
@@ -1,10 +1,9 @@
 """
-Functions to run searches based on Haystack pipelines and print the results. 
+Functions to help clean up input queries and format the results.
 """
 
 import re
-from haystack import Pipeline
-
+from search_backend.api.lib.searchservice import Search
 
 def clean_query(query):
     """
@@ -50,86 +49,6 @@ def detect_bad_query(query):
         return True
 
     return False
-
-
-class Search:
-    """
-    Run different types of search: either BM25, semantic, or hybrid (semantic + BM25)
-    """
-
-    def __init__(self, search_query: str, pipeline: Pipeline, filters: dict=None, top_k: int=10):
-        """
-        Args:
-            :search_query: The search query in the form of a text string.
-            :pipeline: The pipeline to use. This should be defined using the RetrievalPipeline() class.
-            :filters: Metadata filters. These should be formatted like:
-                ```
-                filters = {
-                    "operator": "AND",
-                    "conditions": [
-                        {"field": "meta.type", "operator": "==", "value": "article"},
-                        {"field": "meta.genre", "operator": "in", "value": ["economy", "politics"]},
-                    ],
-                }
-                ```
-            :top_k: How many results to return.
-        """
-
-        self.search_query = search_query
-        self.pipeline = pipeline
-        self.filters = filters
-        self.top_k = top_k
-
-    def hybrid_search(self):
-        """
-        Run a hybrid search pipeline and return results.
-
-        :return: A list of ranked search results.
-        """
-
-        prediction = self.pipeline.run(
-            {
-                "dense_text_embedder": {"text": self.search_query},
-                "bm25_retriever": {"query": self.search_query, "filters": self.filters, "top_k": self.top_k},
-                "embedding_retriever": {"filters": self.filters, "top_k": self.top_k},
-                "ranker": {"query": self.search_query, "top_k": self.top_k},
-            }
-        )
-
-        return prediction
-
-    def semantic_search(self):
-        """
-        Run a semantic search pipeline and return results.
-
-        :return: A list of ranked search results.
-        """
-
-        print("Running search...")
-        prediction = self.pipeline.run(
-            {
-                "dense_text_embedder": {"text": self.search_query},
-                "embedding_retriever": {"filters": self.filters, "top_k": self.top_k},
-                "ranker": {"query": self.search_query, "top_k": self.top_k},
-            }
-        )
-
-        return prediction
-
-    def bm25_search(self):
-        """
-        Run a BM25 search pipeline and return results.
-
-        :return: A list of ranked search results.
-        """
-
-        prediction = self.pipeline.run(
-            {
-                "bm25_retriever": {"query": self.search_query, "filters": self.filters, "top_k": self.top_k},
-            }
-        )
-
-        return prediction
 
 
 def pretty_print_results(prediction):
